CLI config command refactored.
